Update ResourceTest to new pattern
